tests/integration/pyproject.toml
[project]
name = "image-factory-integration-tests"
version = "0.1.0"
description = "Image Factory Integration Tests"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["app"]
